# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AI健身系统API",
    "description": "AI健身系统后端API接口文档",
    "version": "1.0.0",
    "contact": {
      "name": "AI健身系统开发团队",
      "email": "contact@fitness-system.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "本地开发环境"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/plans/generate-from-profile": {
      "post": {
        "summary": "从个人档案生成计划",
        "description": "从个人档案自动生成健身计划",
        "tags": ["健身计划模块"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "生成成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultPlanData"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/plans/generate": {
      "post": {
        "summary": "手动生成计划",
        "description": "手动选择参数生成健身计划",
        "tags": ["健身计划模块"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanGenerateDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "生成成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultLong"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/plans/my": {
      "get": {
        "summary": "获取我的计划列表",
        "description": "获取当前用户的健身计划列表",
        "tags": ["健身计划模块"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultListPlanVO"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/plans/{planId}": {
      "get": {
        "summary": "获取计划详情",
        "description": "获取健身计划详细信息",
        "tags": ["健身计划模块"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "planId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "计划ID"
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultPlanVO"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "删除计划",
        "description": "删除健身计划",
        "tags": ["健身计划模块"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "planId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "计划ID"
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "请在请求头中添加：Authorization: Bearer {token}"
      }
    },
    "schemas": {
      "Result": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "状态码"
          },
          "message": {
            "type": "string",
            "description": "消息"
          },
          "data": {
            "type": "object",
            "description": "数据"
          }
        },
        "example": {
          "code": 200,
          "message": "success",
          "data": {}
        }
      },
...
      "ResultLong": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "integer",
            "format": "int64",
            "description": "ID值"
          }
        }
      },
...
      "ResultListPlanVO": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlanVO"
            }
          }
        }
      },
      "ResultPlanVO": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/PlanVO"
          }
        }
      },
      "ResultPlanData": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "object",
            "properties": {
              "planName": {
                "type": "string",
                "description": "计划名称"
              },
              "weeklyPlan": {
                "type": "object",
                "description": "周计划"
              }
            }
          }
        }
      },
...
      "PlanGenerateDTO": {
        "type": "object",
        "properties": {
          "goal": {
            "type": "string",
            "description": "健身目标"
          },
          "bodyPart": {
            "type": "string",
            "description": "训练部位"
          },
          "experience": {
            "type": "string",
            "description": "经验水平"
          },
          "workoutTime": {
            "type": "string",
            "description": "训练时间"
          },
          "equipment": {
            "type": "string",
            "description": "可用器材"
          }
        },
        "required": ["goal", "bodyPart", "experience", "workoutTime", "equipment"]
      },
...
      "PlanVO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "计划ID"
          },
          "planName": {
            "type": "string",
            "description": "计划名称"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "创建时间"
          },
          "details": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlanDetailVO"
            },
            "description": "计划详情"
          }
        }
      },
      "PlanDetailVO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "详情ID"
          },
          "planId": {
            "type": "integer",
            "description": "计划ID"
          },
          "day": {
            "type": "string",
            "description": "训练日"
          },
          "exercises": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "description": "运动名称"
                },
                "sets": {
                  "type": "integer",
                  "description": "组数"
                },
                "reps": {
                  "type": "integer",
                  "description": "次数"
                },
                "duration": {
                  "type": "integer",
                  "description": "时长（分钟）"
                }
              }
            },
            "description": "运动列表"
          }
        }
      },
...
    }
  }
}